pyproject.toml
[project]
name = "video-engine"
version = "0.1.0"
description = "AI video generation engine — narration, visuals, rendering pipeline"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.84.0",
    "elevenlabs>=2.39.0",
    "openai>=2.28.0",
    "pillow>=12.1.1",
    "faster-whisper>=1.2.1",
    "moviepy>=1.0.3",
    "opencv-python-headless>=4.13.0.92",
    "numpy",
    "requests",
    "structlog>=24.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["video_engine"]
